+++ sdk/java/java/lang/AutoCloseable.java
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-only
-package java.lang;
-
-public interface AutoCloseable {
-  void close() throws Exception;
-}